+++ ecmascript/transforms/src/compat/es2015/regenerator/mod.rs
@@ -1,3 +1,4 @@
+use self::case::CaseHandler;
 use crate::{
     pass::Pass,
     util::{prepend, ExprFactory, StmtLike},
@@ -5,6 +6,8 @@ use crate::{
 use ast::*;
 use std::mem::replace;
 use swc_common::{Fold, FoldWith, Spanned, Visit, VisitWith, DUMMY_SP};
+
+mod case;
 
 pub fn regenerator() -> impl Pass {
     Regenerator::default()
@@ -148,6 +151,7 @@ impl Regenerator {
                     ctx: &ctx,
                     idx: &mut idx,
                     temp_idx: &mut temp_idx,
+                    stmts: &mut case.cons,
                 };
 
                 let stmt = stmt.fold_with(&mut handler);
@@ -279,67 +283,6 @@ impl Regenerator {
                 ..f
             },
         )
-    }
-}
-
-#[derive(Debug)]
-struct CaseHandler<'a> {
-    ctx: &'a Ident,
-    idx: &'a mut u32,
-    temp_idx: &'a mut u32,
-}
-
-impl Fold<Stmt> for CaseHandler<'_> {
-    fn fold(&mut self, s: Stmt) -> Stmt {
-        let s: Stmt = s.fold_children(self);
-
-        match s {
-            Stmt::Return(ret) => {
-                return ReturnStmt {
-                    arg: Some(
-                        box CallExpr {
-                            span: DUMMY_SP,
-                            callee: self.ctx.clone().member(quote_ident!("abrupt")).as_callee(),
-                            args: {
-                                let ret_arg = Lit::Str(Str {
-                                    span: DUMMY_SP,
-                                    value: "return".into(),
-                                    has_escape: false,
-                                })
-                                .as_arg();
-
-                                if let Some(arg) = ret.arg {
-                                    vec![ret_arg, arg.as_arg()]
-                                } else {
-                                    vec![ret_arg]
-                                }
-                            },
-                            type_args: Default::default(),
-                        }
-                        .into(),
-                    ),
-                    ..ret
-                }
-                .into()
-            }
-            _ => {}
-        }
-
-        s
-    }
-}
-
-impl Fold<Function> for CaseHandler<'_> {
-    #[inline(always)]
-    fn fold(&mut self, f: Function) -> Function {
-        f
-    }
-}
-
-impl Fold<ArrowExpr> for CaseHandler<'_> {
-    #[inline(always)]
-    fn fold(&mut self, f: ArrowExpr) -> ArrowExpr {
-        f
     }
 }
